- adds internet connection check in BaseActivity - Updates page offline verification logic to throw an exception on failure - adds a separate suite for offline test
