+++ app/src/androidTest/java/org/wikipedia/base/BaseTest.kt
@@ -91,6 +91,10 @@ abstract class BaseTest<T : AppCompatActivity>(
         Thread.sleep(TestConfig.DELAY_MEDIUM)
     }
 
+    fun isOnline(): Boolean {
+        return WikipediaApp.instance.isOnline
+    }
+
     @After
     open fun tearDown() {
         Intents.release()
