+++ app/src/androidTest/java/org/wikipedia/testsuites/ArticlesTestSuite.kt
@@ -10,10 +10,10 @@ import org.wikipedia.tests.articles.ArticleTabTest
 import org.wikipedia.tests.articles.EditIconTest
 import org.wikipedia.tests.articles.LeadNonLeadImageAndPreviewLinkTest
 import org.wikipedia.tests.articles.MediaTest
-import org.wikipedia.tests.articles.SavedArticleOnlineOfflineTest
 import org.wikipedia.tests.articles.SavedArticleTest
 import org.wikipedia.tests.articles.SpecialArticleTest
 import org.wikipedia.tests.articles.TableOfContentsTest
+import org.wikipedia.tests.offline.SavedArticleOnlineOfflineTest
 
 @RunWith(Suite::class)
 @SuiteClasses(
